requirements.txt
django>=4.2.11, <5.0
djangorestframework>=3.15.1,<4.0
channels==4.0.0
django-environ==0.11.2
psycopg2-binary==2.9.9
celery==5.3.6
redis==5.0.3
django-cors-headers==4.3.1
requests==2.31.0
Authlib==1.3.0
requests-oauthlib==2.0.0
django-filter==24.2